# backlot-agentic-cinema/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ToolField {
  name: string;
  type: string;
  description: string;
}

export interface ToolAction {
  label: string;
  description: string;
}

export type ToolSpecMode = typeof ToolSpecMode[keyof typeof ToolSpecMode];


export const ToolSpecMode = {
  production: 'production',
  interactive: 'interactive',
} as const;

export interface GroundedRecord {
  sourceType: string;
  name: string;
  details: string;
}

export type InteractiveSpecMechanic = typeof InteractiveSpecMechanic[keyof typeof InteractiveSpecMechanic];


export const InteractiveSpecMechanic = {
  match: 'match',
  branch: 'branch',
} as const;

export interface InteractiveSpec {
  mechanic: InteractiveSpecMechanic;
  entities: GroundedRecord[];
  winCondition: string;
  instructions: string;
}

export interface ToolSpec {
  name: string;
  mode: ToolSpecMode;
  purpose: string;
  fields: ToolField[];
  actions: ToolAction[];
  initialData: GroundedRecord[];
  interactive: InteractiveSpec | null;
}

export interface GroundedFacts {
  characters: string[];
  scenes: string[];
  props: string[];
  locations: string[];
  rightsMentions: string[];
}

export type ToolRequestStatus = typeof ToolRequestStatus[keyof typeof ToolRequestStatus];


export const ToolRequestStatus = {
  pending: 'pending',
  'facts-ready': 'facts-ready',
  'spec-ready': 'spec-ready',
  'validation-failed': 'validation-failed',
  built: 'built',
} as const;

/**
 * @nullable
 */
export type ToolRequestSupervisorStatus = typeof ToolRequestSupervisorStatus[keyof typeof ToolRequestSupervisorStatus] | null;


export const ToolRequestSupervisorStatus = {
  passed: 'passed',
  failed: 'failed',
} as const;

export interface ToolRequest {
  id: number;
  requestText: string;
  spec: ToolSpec | null;
  groundedFacts: GroundedFacts | null;
  /** @nullable */
  sourcePdfName: string | null;
  status: ToolRequestStatus;
  createdAt: string;
  /** @nullable */
  generatedCode: string | null;
  /** @nullable */
  supervisorStatus: ToolRequestSupervisorStatus;
  /** @nullable */
  supervisorReason: string | null;
  /** @nullable */
  publishedUrl: string | null;
  /** @nullable */
  publishedSlug: string | null;
}

export interface ToolRequestInput {
  /** @minLength 3 */
  requestText: string;
}

export interface PdfGroundingInput {
  /** @minLength 3 */
  requestText: string;
  /** @minLength 20 */
  pdfBase64: string;
  /** @minLength 1 */
  pdfName: string;
}

export type GroundedSpecInputMode = typeof GroundedSpecInputMode[keyof typeof GroundedSpecInputMode];


export const GroundedSpecInputMode = {
  production: 'production',
  interactive: 'interactive',
} as const;

export interface GroundedSpecInput {
  mode: GroundedSpecInputMode;
}

export interface Error {
  error: string;
}