# backlot-agentic-cinema/lib/api-zod/src/generated/types/groundedSpecInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GroundedSpecInputMode } from './groundedSpecInputMode';

export interface GroundedSpecInput {
  mode: GroundedSpecInputMode;
}