# backlot-agentic-cinema/lib/api-zod/src/generated/types/interactiveSpec.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GroundedRecord } from './groundedRecord';
import type { InteractiveSpecMechanic } from './interactiveSpecMechanic';

export interface InteractiveSpec {
  mechanic: InteractiveSpecMechanic;
  entities: GroundedRecord[];
  winCondition: string;
  instructions: string;
}