# backlot-agentic-cinema/lib/api-zod/src/generated/types/pdfGroundingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PdfGroundingInput {
  /** @minLength 3 */
  requestText: string;
  /** @minLength 20 */
  pdfBase64: string;
  /** @minLength 1 */
  pdfName: string;
}