# backlot-agentic-cinema/lib/api-zod/src/generated/types/toolAction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ToolAction {
  label: string;
  description: string;
}