# backlot-agentic-cinema/lib/api-zod/src/generated/types/toolSpec.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GroundedRecord } from './groundedRecord';
import type { InteractiveSpec } from './interactiveSpec';
import type { ToolAction } from './toolAction';
import type { ToolField } from './toolField';
import type { ToolSpecMode } from './toolSpecMode';

export interface ToolSpec {
  name: string;
  mode: ToolSpecMode;
  purpose: string;
  fields: ToolField[];
  actions: ToolAction[];
  initialData: GroundedRecord[];
  interactive: InteractiveSpec | null;
}